# src/core/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Mini E-Commerce API
  description: API for managing users, products, cart operations, orders, and fraud detection.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDTO'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      summary: User login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDTO'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/profile:
    get:
      summary: Get current user profile
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Products
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          example:
            code: 400
            message: Bad Request
            fields:
              - field: field1
                values:
                  - field1 cannot be empty.
                  - field1 cannot be empty.

              - field: field2
                values:
                  - field2 must be at least 8 characters long.
                  - field2 must be at least 8 characters long.
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          example:
            code: 401
            statusCode: Unauthorized
...
  schemas:
...
    RegisterDTO:
      type: object
      required:
        - email
        - password
        - fullname
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        fullName:
          type: string
    LoginDTO:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    # Entities
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, restricted]
        role:
          type: string
          enum: [admin, customer]
        createdAt:
          type: string
          format: date-time
...